feat: Add TodoInput and Todos components
This commit adds the TodoInput and Todos components to the project. The TodoInput component is responsible for rendering an input field and an icon for adding new tasks. The Todos component is a placeholder component that will display the list of tasks.

The changes include:
- Importing the LuCircle icon from the react-icons/lu package in TodoInput.jsx
- Updating the JSX code in TodoInput.jsx to include the input field and icon
- Importing the Todos component in TodoList.jsx
- Adding the Todos component to the JSX code in TodoList.jsx
- Creating a new file, Todos.jsx, which contains the Todos component

These changes are necessary to implement the basic functionality of adding and displaying tasks in the TodoList component.